prevent errors when starting or stopping a measure
To ease usage, I've removed the error throwing for this two cases:
1. on start if a measure with the same name is already running
2. on stop if the measure never started

For case 1. there is no need to throw an error as also Timing API
allow multiple markers with the same name; the only caveat is that
computing the measure only the last one will be considered.

For case 2., we are now returning `FALSE` as a way to check success.
Again no need to throw an error has nothing bad will happen.